+++ lib/express/operations/poolboy_configs.ex
@@ -1,4 +1,12 @@
 defmodule Express.Operations.PoolboyConfigs do
+  @moduledoc "Returns poolboy configurations."
+
+  @doc """
+  Returns poolboy configuration for the Buffer adders.
+  If configuration described in config file - returns it.
+  Returns default configuration otherwise.
+  """
+  @spec buffer_adders() :: %{config: Keyword.t, name: atom()}
   def buffer_adders do
     %{
       config: buffer_adders_pool_config(),
@@ -6,6 +14,12 @@ defmodule Express.Operations.PoolboyConfigs do
     }
   end
 
+  @doc """
+  Returns poolboy configuration for the FCM workers.
+  If configuration described in config file - returns it.
+  Returns default configuration otherwise.
+  """
+  @spec fcm_workers() :: %{config: Keyword.t, name: atom()}
   def fcm_workers do
     %{
       config: fcm_workers_pool_config(),
@@ -13,6 +27,12 @@ defmodule Express.Operations.PoolboyConfigs do
     }
   end
 
+  @doc """
+  Returns poolboy configuration for the APNS workers.
+  If configuration described in config file - returns it.
+  Returns default configuration otherwise.
+  """
+  @spec apns_workers() :: %{config: Keyword.t, name: atom()}
   def apns_workers do
     %{
       config: apns_workers_pool_config(),
@@ -20,6 +40,7 @@ defmodule Express.Operations.PoolboyConfigs do
     }
   end
 
+  @spec buffer_adders_pool_config() :: Keyword.t
   defp buffer_adders_pool_config do
     Application.get_env(:express, :buffer)[:adders_pool_config] ||
     [
@@ -30,11 +51,13 @@ defmodule Express.Operations.PoolboyConfigs do
     ]
   end
 
+  @spec buffer_adders_pool_name() :: atom()
   defp buffer_adders_pool_name do
     [{:name, {_, name}} | _] = buffer_adders_pool_config()
     name
   end
 
+  @spec fcm_workers_pool_config() :: Keyword.t
   defp fcm_workers_pool_config do
     Application.get_env(:express, :fcm)[:workers_pool_config] ||
     [
@@ -44,11 +67,13 @@ defmodule Express.Operations.PoolboyConfigs do
     ]
   end
 
+  @spec fcm_workers_pool_name() :: atom()
   defp fcm_workers_pool_name do
     [{:name, {_, name}} | _] = fcm_workers_pool_config()
     name
   end
 
+  @spec apns_workers_pool_config() :: Keyword.t
   defp apns_workers_pool_config do
     Application.get_env(:express, :apns)[:workers_pool_config] ||
     [
@@ -58,6 +83,7 @@ defmodule Express.Operations.PoolboyConfigs do
     ]
   end
 
+  @spec apns_workers_pool_name() :: atom()
   defp apns_workers_pool_name do
     [{:name, {_, name}} | _] = apns_workers_pool_config()
     name