+++ lib/express/push_requests/consumers_supervisor.ex
@@ -1,4 +1,8 @@
 defmodule Express.PushRequests.ConsumersSupervisor do
+  @moduledoc """
+  Dynamically spawns and supervises consumers for the push requests buffer.
+  """
+
   use Supervisor
 
   def start_link do
@@ -21,6 +25,8 @@ defmodule Express.PushRequests.ConsumersSupervisor do
     [strategy: :simple_one_for_one, name: __MODULE__]
   end
 
+  @doc "Spawns the buffer consumer."
+  @spec start_consumer() :: Supervisor.on_start_child
   def start_consumer do
     Supervisor.start_child(__MODULE__, [])
   end