+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## [1.1.0] - 2017.10.26
+
+* JWT for APNS
+* GenStage for load balancing
+* Refactored supervision tree
+
 ## [1.0.0] - 2017.07.23
 
 the first publication on hex.pm